Input and output embeddings fully connected with
I have connected input and output bert embedding to the transformer model. No my next step is to make sure that code is working fine and make py files .After the the generator module will be almost fully complete. In the current case, only linear layer and gumbell activation needs to be applied